requirements.txt
pandas
polygon-api-client
numpy
http.client 
urllib
os
sqlalchemy
